add sweat alert for delete

+++ js/main.js
@@ -70,11 +70,29 @@ btnThemNV = () => {
 };
 
 deleteNv = (account) => {
-  for (let i = 0; i < danhSach.length; i++) {
-    if (danhSach[i].account == account) {
-      danhSach.splice(i, 1);
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!',
+  }).then((result) => {
+    if (result.isConfirmed) {
+      for (let i = 0; i < danhSach.length; i++) {
+        if (danhSach[i].account == account) {
+          danhSach.splice(i, 1);
+        }
+      }
+      localStorage.setItem('danhSach', JSON.stringify(danhSach));
+      Swal.fire('Deleted!', 'Your file has been deleted.', 'success').then(
+        (result) => {
+          if (result.isConfirmed) {
+            window.location.reload();
+          }
+        }
+      );
     }
-  }
-  localStorage.setItem('danhSach', JSON.stringify(danhSach));
-  location.reload();
+  });
 };
